GPU_Deploy_Report/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
fastapi>=0.104.0
uvicorn>=0.24.0
pillow>=10.0.0
python-multipart>=0.0.6
